validators: creates validate_objects_filenames validator
This is a workaround to validate what is not possible to do with
pure JSON Schema. This validator prevents users to create a metadata
with objects sets with different files.

To be clear, by now, we only allow users to have different
configurations between sets, but they must have the same filenames.

Signed-off-by: Pablo Palácios <ppalacios992@gmail.com>